重新调整了文件布局。

+++ src/StreamWrapper.php
@@ -7,6 +7,9 @@ use InvalidArgumentException;
 use Psr\Http\Message\StreamInterface;
 use fize\io\Stream;
 
+/**
+ * 流包装器
+ */
 class StreamWrapper
 {
 
@@ -25,6 +28,26 @@ class StreamWrapper
      */
     private $mode;
 
+    /**
+     * @var string 路径
+     */
+    protected $path;
+
+    /**
+     * @var array 选项
+     */
+    protected $options;
+
+    /**
+     * @var int 标识
+     */
+    protected $flags;
+
+    /**
+     * 获取流资源
+     * @param StreamInterface $stream 流对象
+     * @return resource 失败时返回false
+     */
     public static function getResource(StreamInterface $stream)
     {
         self::register();
@@ -37,7 +60,7 @@ class StreamWrapper
             throw new InvalidArgumentException('The stream must be readable, writable, or both.');
         }
 
-        return fopen('guzzle://stream', $mode, null, self::createStreamContext($stream));
+        return fopen('fize://stream', $mode, null, self::createStreamContext($stream));
     }
 
     /**
@@ -45,57 +68,91 @@ class StreamWrapper
      */
     public static function register()
     {
-        if (!in_array('guzzle', Stream::getwrappers())) {
-            Stream::wrapperRegister('guzzle', __CLASS__);
+        if (!in_array('fize', Stream::getwrappers())) {
+            Stream::wrapperRegister('fize', __CLASS__);
         }
     }
 
     /**
      * 创建流上下文
-     * @param StreamInterface $stream 流
+     * @param StreamInterface $stream 流对象
      * @return resource
      */
     public static function createStreamContext(StreamInterface $stream)
     {
         return Stream::contextCreate([
-            'guzzle' => ['stream' => $stream]
+            'fize' => ['stream' => $stream]
         ]);
     }
 
+    /**
+     * 打开文件或者URL
+     * @param string $path 文件路径或者URL
+     * @param string $mode 模式
+     * @param array $options 选项
+     * @param string $opened_path
+     * @return bool 如果路径被成功打开，该值返回实际路径
+     */
     public function streamOpen($path, $mode, $options, &$opened_path)
     {
+        $this->path = $path;
+        $this->options = $options;
         $options = stream_context_get_options($this->context);
 
-        if (!isset($options['guzzle']['stream'])) {
+        if (!isset($options['fize']['stream'])) {
             return false;
         }
 
         $this->mode = $mode;
-        $this->stream = $options['guzzle']['stream'];
-
+        $this->stream = $options['fize']['stream'];
+        $opened_path = realpath($path);
         return true;
     }
 
+    /**
+     * 读取
+     * @param int $count 字节数
+     * @return string
+     */
     public function streamRead($count)
     {
         return $this->stream->read($count);
     }
 
+    /**
+     * 写入
+     * @param string $data 数据
+     * @return int
+     */
     public function streamWrite($data)
     {
         return (int)$this->stream->write($data);
     }
 
+    /**
+     * 返回当前流位置
+     * @return int
+     */
     public function streamTell()
     {
         return $this->stream->tell();
     }
 
+    /**
+     * 是否到流的结尾
+     * @return bool
+     */
     public function streamEof()
     {
         return $this->stream->eof();
     }
 
+    /**
+     * 移动流位置
+     * @param int $offset 偏移
+     * @param int $whence 偏移参照
+     * @return bool
+     */
     public function streamSeek($offset, $whence)
     {
         $this->stream->seek($offset, $whence);
@@ -103,13 +160,25 @@ class StreamWrapper
         return true;
     }
 
+    /**
+     * 返回底层资源
+     * @param int $cast_as STREAM_CAST_FOR_SELECT|STREAM_CAST_AS_STREAM
+     * @return resource|null
+     */
     public function streamCast($cast_as)
     {
-        $stream = clone($this->stream);
+        if ($cast_as == STREAM_CAST_FOR_SELECT) {
+            return null;
+        }
 
+        $stream = clone($this->stream);
         return $stream->detach();
     }
 
+    /**
+     * 检索关于文件资源的信息
+     * @return array
+     */
     public function streamStat()
     {
         static $modeMap = [
@@ -137,8 +206,17 @@ class StreamWrapper
         ];
     }
 
+    /**
+     * 检索关于文件的信息
+     * @param string $path 路径或者URL
+     * @param int $flags 标识
+     * @return array
+     */
     public function urlStat($path, $flags)
     {
+        $this->path = $path;
+        $this->flags = $flags;
+
         return [
             'dev'     => 0,
             'ino'     => 0,
